Fix minor bugs about value type.
Both requestId and timestamp is defined as 'long' in aidl, which
is 'int64_t' in c++.

Test: Presubmit
Bug: 201830716

+++ automotive/vehicle/aidl/impl/hardware/include/IVehicleHardware.h
@@ -39,7 +39,7 @@ struct DumpResult {
 // A request type for 'setValues' or 'getValues' method.
 struct VehiclePropValueRequest {
     // A unique request ID set by the sender.
-    int requestId;
+    int64_t requestId;
     // The property to get/set.
     ::aidl::android::hardware::automotive::vehicle::VehiclePropValue value;
 };
